Refactor MockGameModule
